Update ColorList style
