+++ ODS.swift
@@ -1,0 +1,41 @@
+//
+//  ODS.swift
+//  OrangeDesignSystem
+//
+//  Created by Guillaume Marge on 25/03/2022.
+//
+
+import Foundation
+
+// shortcuts : add fonts ?
+public enum ODS {
+    public static let coreOrange = ODSColor.coreOrange.color
+    public static let coreWhite = ODSColor.coreWhite.color
+    public static let coreBlack = ODSColor.coreBlack.color
+    public static let coreObsGrey = ODSColor.coreObsGrey.color
+
+    public static let functionalPositive = ODSColor.functionalPositive.color
+    public static let functionalNegative = ODSColor.functionalNegative.color
+    public static let functionalAlert = ODSColor.functionalAlert.color
+    public static let functionalInfo = ODSColor.functionalInfo.color
+
+    public static let supportingBlue100 = ODSColor.supportingBlue100.color
+    public static let supportingBlue200 = ODSColor.supportingBlue200.color
+    public static let supportingBlue300 = ODSColor.supportingBlue300.color
+
+    public static let supportingGreen100 = ODSColor.supportingGreen100.color
+    public static let supportingGreen200 = ODSColor.supportingGreen200.color
+    public static let supportingGreen300 = ODSColor.supportingGreen300.color
+
+    public static let supportingYellow100 = ODSColor.supportingYellow100.color
+    public static let supportingYellow200 = ODSColor.supportingYellow200.color
+    public static let supportingYellow300 = ODSColor.supportingYellow300.color
+
+    public static let supportingPurple100 = ODSColor.supportingPurple100.color
+    public static let supportingPurple200 = ODSColor.supportingPurple200.color
+    public static let supportingPurple300 = ODSColor.supportingPurple300.color
+
+    public static let supportingPink100 = ODSColor.supportingPink100.color
+    public static let supportingPink200 = ODSColor.supportingPink200.color
+    public static let supportingPink300 = ODSColor.supportingPink300.color
+}
