+++ OrangeDesignSystemDemo/Guidelines/ColorsList.swift
@@ -35,65 +35,73 @@ struct ColorList: View {
     // MARK: - Body
     var body: some View {
         VStack {
+            Text("Palette").odsFont(style: .title1).frame(maxWidth: .infinity, alignment: .leading).padding(EdgeInsets(top: 0, leading: 15, bottom: 5, trailing: 15))
+
             Picker("Favorite Scheme", selection: $screenColorScheme, content: {
                 Text("On Light").tag(ColorScheme.light)
                 Text("On Dark").tag(ColorScheme.dark)
             }).pickerStyle(.segmented)
                 .onAppear {
                     screenColorScheme = phoneColorScheme
-                }.padding(16)
+                }.padding(EdgeInsets(top: 10, leading: 15, bottom: 5, trailing: 15))
+
             ScrollView {
+                Spacer().frame(height: 5)
                 SectionTitle(title: "Core", scheme: $screenColorScheme)
                 VStack {
-                    HStack(spacing: 19) {
+                    Spacer().frame(height: 30)
+                    HStack(spacing: 15) {
                         colorBigView(color: ODSColor.coreOrange, scheme: $screenColorScheme, bordered: false)
                         colorBigView(color: ODSColor.coreWhite, scheme: $screenColorScheme, bordered: screenColorScheme == .light)
                     }
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorBigView(color: ODSColor.coreBlack, scheme: $screenColorScheme, bordered: screenColorScheme == .dark)
                         colorBigView(color: ODSColor.coreObsGrey, scheme: $screenColorScheme, bordered: false)
                     }
                 }
+                Spacer().frame(height: 30)
                 SectionTitle(title: "Functional", scheme: $screenColorScheme)
                 VStack {
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorBigView(color: ODSColor.functionalPositive, scheme: $screenColorScheme, bordered: false)
                         colorBigView(color: ODSColor.functionalNegative, scheme: $screenColorScheme, bordered: false)
                     }
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorBigView(color: ODSColor.functionalAlert, scheme: $screenColorScheme, bordered: false)
                         colorBigView(color: ODSColor.functionalInfo, scheme: $screenColorScheme, bordered: false)
                     }
                 }
+                Spacer().frame(height: 30)
                 SectionTitle(title: "Supporting", scheme: $screenColorScheme)
                 VStack {
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorSmallView(color: ODSColor.supportingBlue100, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingBlue200, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingBlue300, scheme: $screenColorScheme)
                     }
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorSmallView(color: ODSColor.supportingYellow100, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingYellow200, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingYellow300, scheme: $screenColorScheme)
                     }
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorSmallView(color: ODSColor.supportingGreen100, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingGreen200, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingGreen300, scheme: $screenColorScheme)
                     }
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorSmallView(color: ODSColor.supportingPink100, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingPink200, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingPink300, scheme: $screenColorScheme)
                     }
-                    HStack(spacing: 19) {
+                    HStack(spacing: 15) {
                         colorSmallView(color: ODSColor.supportingPurple100, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingPurple200, scheme: $screenColorScheme)
                         colorSmallView(color: ODSColor.supportingPurple300, scheme: $screenColorScheme)
                     }
                 }
-            }.padding(16).background(screenColorScheme == .light ? Color.white : Color.black)
+            }.padding(EdgeInsets(top: 0, leading: 15, bottom: 5, trailing: 15))
+                .background(screenColorScheme == .light ? Color.white : Color.black)
         }
     }
 }
@@ -103,8 +111,7 @@ struct SectionTitle: View {
     @Binding var scheme: ColorScheme
 
     var body: some View {
-        Spacer().frame(height: 30)
-        Text(title).odsFont(style: .title2)
+        Text(title).odsFont(style: .title1)
             .foregroundColor(scheme == .dark ? Color.white : Color.black)
             .frame(maxWidth: .infinity, alignment: .leading)
     }
